Move method body generation to generators folder
