+++ kybra/compiler/kybra_generate/src/py_ast/kybra_types/kybra_stmt/canister_method.rs
@@ -1,14 +1,11 @@
-use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
-use rustpython_parser::ast::{ExprKind, StmtKind};
-
-use crate::{generators::tuple, py_ast::kybra_types::KybraExpr};
 use cdk_framework::{
     nodes::{data_type_nodes::ActPrimitiveLit, ActFnParam},
     ActDataType, CanisterMethod, CanisterMethodType, ToActDataType,
 };
+use rustpython_parser::ast::{ExprKind, StmtKind};
 
 use super::KybraStmt;
+use crate::{generators::canister_methods::query_and_update, py_ast::kybra_types::KybraExpr};
 
 impl KybraStmt<'_> {
     pub fn is_canister_method_stmt(&self) -> bool {
@@ -42,7 +39,7 @@ impl KybraStmt<'_> {
         }
     }
 
-    fn build_act_params(&self) -> Vec<ActFnParam> {
+    pub fn build_act_params(&self) -> Vec<ActFnParam> {
         match &self.stmt_kind.node {
             StmtKind::FunctionDef { args, .. } => {
                 args.args
@@ -64,7 +61,7 @@ impl KybraStmt<'_> {
         }
     }
 
-    fn is_manual(&self) -> bool {
+    pub fn is_manual(&self) -> bool {
         match &self.stmt_kind.node {
             StmtKind::FunctionDef { returns, .. } => match returns {
                 Some(returns) => KybraExpr {
@@ -99,7 +96,7 @@ impl KybraStmt<'_> {
     pub fn as_canister_method(&self) -> Option<CanisterMethod> {
         match &self.stmt_kind.node {
             StmtKind::FunctionDef { name, .. } => {
-                let body = self.generate_body();
+                let body = query_and_update::generate_body(&self);
                 let params = self.build_act_params();
                 let return_type = self.build_act_return_type();
 
@@ -117,7 +114,7 @@ impl KybraStmt<'_> {
         }
     }
 
-    fn build_act_return_type(&self) -> ActDataType {
+    pub fn build_act_return_type(&self) -> ActDataType {
         let returns = match &self.stmt_kind.node {
             StmtKind::FunctionDef { returns, .. } => returns,
             _ => panic!("Unreachable"),
@@ -133,95 +130,5 @@ impl KybraStmt<'_> {
             }
             None => ActPrimitiveLit::Void.to_act_data_type(&None),
         }
-    }
-
-    fn generate_body(&self) -> TokenStream {
-        let act_params = self.build_act_params();
-
-        let name = match self.get_name() {
-            Some(name) => name,
-            None => todo!(),
-        };
-
-        let param_conversions = act_params
-            .iter()
-            .map(|param| {
-                let name = format_ident!("{}", param.prefixed_name());
-                quote! {
-                    #name.try_into_vm_value(vm).unwrap()
-                }
-            })
-            .collect();
-
-        let params = tuple::generate_tuple(&param_conversions);
-
-        let return_expression = self.generate_return_expression();
-
-        quote! {
-            unsafe {
-                let _kybra_interpreter = _KYBRA_INTERPRETER_OPTION.as_mut().unwrap();
-                let _kybra_scope = _KYBRA_SCOPE_OPTION.as_mut().unwrap();
-
-                let vm = &_kybra_interpreter.vm;
-
-                let method_py_object_ref = _kybra_unwrap_rust_python_result(_kybra_scope.globals.get_item(#name, vm), vm);
-
-                let invoke_result = vm.invoke(&method_py_object_ref, #params);
-
-                match invoke_result {
-                    Ok(py_object_ref) => {
-                        let _kybra_final_return_value = _kybra_async_result_handler(vm, &py_object_ref, vm.ctx.none()).await;
-
-                        #return_expression
-                    },
-                    Err(err) => {
-                        let err_string: String = err.to_pyobject(vm).repr(vm).unwrap().to_string();
-
-                        panic!("{}", err_string);
-                    }
-                }
-            }
-        }
-    }
-
-    fn generate_return_expression(&self) -> TokenStream {
-        if self.is_manual() {
-            return quote! {
-                ic_cdk::api::call::ManualReply::empty()
-            };
-        }
-
-        let return_type = self.build_act_return_type();
-        if type_is_null_or_void(return_type) {
-            return quote! {
-                return;
-            };
-        }
-
-        quote! {
-            _kybra_final_return_value.try_from_vm_value(vm).unwrap()
-        }
-    }
-}
-
-fn type_is_null_or_void(act_type: ActDataType) -> bool {
-    match act_type {
-        ActDataType::Primitive(primitive) => match primitive.act_type {
-            cdk_framework::nodes::data_type_nodes::LiteralOrTypeAlias::Literal(literal) => {
-                match literal {
-                    ActPrimitiveLit::Null => true,
-                    ActPrimitiveLit::Void => true,
-                    _ => false,
-                }
-            }
-            cdk_framework::nodes::data_type_nodes::LiteralOrTypeAlias::TypeAlias(type_alias) => {
-                match type_alias.aliased_type {
-                    ActPrimitiveLit::Null => true,
-                    ActPrimitiveLit::Void => true,
-                    _ => false,
-                }
-            }
-        },
-        _ => false,
     }
 }
